Split MooseMeshUtils into MeshTriangulationUtils and BoundaryLayerUtils
Move the Delaunay triangulation core (XYDelaunayOptions, triangulateWithDelaunay)
into a new MeshTriangulationUtils namespace, and the boundary-layer ring builder
plus its polyline-offset helpers (buildBoundaryLayerRing, generateOffsetPolyline,
collectKeyPointsFromMesh, getKeyNormal) into a new BoundaryLayerUtils namespace.
Update XYDelaunayGenerator, XYTriangleBoundaryLayerGenerator, and the reactor
module's GapLineMeshGenerator to use the new namespaces. buildPolyLineMesh and
other general helpers stay in MooseMeshUtils.
