+++ framework/src/utils/MooseMeshUtils.C
@@ -9,7 +9,6 @@
 
 // MOOSE includes
 #include "MooseMeshUtils.h"
-#include "CastUniquePointer.h"
 
 #include "libmesh/elem.h"
 #include "libmesh/boundary_info.h"
@@ -26,15 +25,7 @@
 #include "libmesh/mesh_communication.h"
 #include "libmesh/edge_edge3.h"
 #include "libmesh/enum_to_string.h"
-#include "libmesh/fe_base.h"
-#include "libmesh/mesh_modification.h"
-#include "libmesh/mesh_serializer.h"
-#include "libmesh/mesh_triangle_holes.h"
-#include "libmesh/parsed_function.h"
-#include "libmesh/poly2tri_triangulator.h"
 #include "libmesh/unstructured_mesh.h"
-
-#include "GeometryUtils.h"
 
 #include "timpi/parallel_sync.h"
 
@@ -1565,769 +1556,5 @@ addExternalBoundary(MeshBase & mesh, const BoundaryID extern_bid, bool & has_ext
         has_external_bid = true;
         binfo.add_side(elem, i_side, extern_bid);
       }
-}
-
-std::unique_ptr<MeshBase>
-buildBoundaryLayerRing(MeshGenerator & mg,
-                       MeshBase & input_mesh,
-                       const std::vector<BoundaryName> & boundary_names,
-                       unsigned int num_layers,
-                       Real thickness,
-                       Real layer_bias,
-                       bool outward,
-                       const MooseEnum & tri_elem_type,
-                       SubdomainID output_subdomain_id,
-                       const SubdomainName & output_subdomain_name)
-{
-  // Extract seed boundary polyline points (vertices + optional midpoints) from input_mesh.
-  std::set<std::size_t> bdry_id_set;
-  if (!boundary_names.empty())
-  {
-    auto ids = MooseMeshUtils::getBoundaryIDs(input_mesh, boundary_names, true);
-    bdry_id_set.insert(ids.begin(), ids.end());
-  }
-  TriangulatorInterface::MeshedHole bdry_mh(input_mesh, bdry_id_set);
-  std::vector<Point> cur_pts;
-  std::vector<Point> cur_mids;
-  // Preserve every input boundary node (including colinear interior nodes on straight edges) so
-  // the ring's innermost polyline can be stitched back to the input mesh's boundary exactly when
-  // keep_input is requested by a downstream caller.
-  MooseMeshUtils::collectKeyPointsFromMesh(bdry_mh,
-                                           cur_pts,
-                                           cur_mids,
-                                           /*skip_node_reduction=*/true);
-
-  // Geometric progression of incremental thicknesses. layer_thicknesses[i] is the offset
-  // distance from polyline i-1 to polyline i (for i >= 1); layer_thicknesses[0] is unused.
-  std::vector<Real> layer_thicknesses(num_layers + 1);
-  const Real unit_thickness =
-      (layer_bias == 1.0)
-          ? (thickness / num_layers)
-          : (thickness / (std::pow(layer_bias, num_layers) - 1.0) * (layer_bias - 1.0));
-  layer_thicknesses[0] = 0.0;
-  for (auto i : make_range(std::vector<Real>::size_type(1), layer_thicknesses.size()))
-    layer_thicknesses[i] = unit_thickness * std::pow(layer_bias, i - 1);
-
-  // Generate the N+1 polyline meshes. Polylines are indexed so that polyline 0 is geometrically
-  // innermost (smallest) and polyline N is outermost (largest). Iteration order depends on
-  // direction: for outward, build polyline 0 first (= input boundary); for inward, build polyline
-  // N first.
-  std::vector<std::unique_ptr<MeshBase>> polylines(num_layers + 1);
-  for (auto layer_i : make_range(num_layers + 1))
-  {
-    const unsigned int layer_index = outward ? layer_i : (num_layers - layer_i);
-
-    // Build the polyline mesh for storage as polylines[layer_index].
-    auto ply = std::make_unique<ReplicatedMesh>(mg.comm());
-    MooseMeshUtils::buildPolyLineMesh(*ply,
-                                      cur_pts,
-                                      cur_mids,
-                                      /*loop=*/true,
-                                      BoundaryName(),
-                                      BoundaryName(),
-                                      std::vector<unsigned int>({1}));
-    polylines[layer_index] = std::move(ply);
-
-    if (layer_i + 1 < num_layers + 1)
-    {
-      // Build a sacrificial polyline mesh to feed generateOffsetPolyline (it triangulates the
-      // input mesh in place to compute side normals; we discard it afterwards).
-      auto ply_for_offset = std::make_unique<ReplicatedMesh>(mg.comm());
-      MooseMeshUtils::buildPolyLineMesh(*ply_for_offset,
-                                        cur_pts,
-                                        cur_mids,
-                                        /*loop=*/true,
-                                        BoundaryName(),
-                                        BoundaryName(),
-                                        std::vector<unsigned int>({1}));
-      std::unique_ptr<UnstructuredMesh> ply_for_offset_u =
-          dynamic_pointer_cast<UnstructuredMesh>(std::move(ply_for_offset));
-
-      std::vector<Point> next_combined = MooseMeshUtils::generateOffsetPolyline(
-          &mg, ply_for_offset_u, cur_pts, cur_mids, outward, layer_thicknesses[layer_i + 1]);
-      if (cur_mids.empty())
-        cur_pts = std::move(next_combined);
-      else
-      {
-        const auto n_vert = cur_pts.size();
-        cur_pts.assign(next_combined.begin(), next_combined.begin() + n_vert);
-        cur_mids.assign(next_combined.begin() + n_vert, next_combined.end());
-      }
-    }
-  }
-
-  // Triangulate each annulus (between polylines[i] and polylines[i+1]). Stitch each annulus to
-  // the accumulating ring (except the first one).
-  std::unique_ptr<MeshBase> ring;
-  for (auto i : make_range(num_layers))
-  {
-    XYDelaunayOptions xyd_opts;
-    xyd_opts.refine_bdy = false;
-    xyd_opts.verify_holes = false;
-    xyd_opts.stitch_holes = {i > 0};
-    xyd_opts.refine_holes = {false};
-    xyd_opts.tri_elem_type = std::string(tri_elem_type);
-    if (output_subdomain_id != 0)
-    {
-      xyd_opts.has_output_subdomain_id = true;
-      xyd_opts.output_subdomain_id = output_subdomain_id;
-    }
-    if (output_subdomain_name.size())
-    {
-      xyd_opts.has_output_subdomain_name = true;
-      xyd_opts.output_subdomain_name = output_subdomain_name;
-    }
-
-    std::vector<std::unique_ptr<MeshBase>> holes;
-    holes.reserve(1);
-    if (i == 0)
-      holes.push_back(std::move(polylines[0]));
-    else
-      holes.push_back(std::move(ring));
-
-    ring = MooseMeshUtils::triangulateWithDelaunay(
-        mg, std::move(polylines[i + 1]), std::move(holes), xyd_opts);
-  }
-  // We now have 2 * num_layers boundaries
-  // Let's only keep the innermost (1) and outermost (2 * num_layers) boundaries, and remove all
-  // intermediate ring bcids
-  std::vector<BoundaryID> bids_to_delete;
-  for (const auto b : make_range(2 * num_layers))
-  {
-    if (b == 1 || b == (num_layers - 1) * 2)
-      continue;
-    bids_to_delete.push_back(b);
-  }
-  auto & bi = ring->get_boundary_info();
-  for (auto b : bids_to_delete)
-    bi.remove_id(b);
-
-  return ring;
-}
-
-std::unique_ptr<MeshBase>
-triangulateWithDelaunay(MeshGenerator & mg,
-                        std::unique_ptr<MeshBase> boundary_mesh,
-                        std::vector<std::unique_ptr<MeshBase>> hole_meshes,
-                        const XYDelaunayOptions & xyd_opts)
-{
-  // Put the boundary mesh in a local pointer
-  std::unique_ptr<UnstructuredMesh> mesh =
-      dynamic_pointer_cast<UnstructuredMesh>(std::move(boundary_mesh));
-
-  // Get ready to triangulate the line segments we extract from it
-  libMesh::Poly2TriTriangulator poly2tri(*mesh);
-  poly2tri.triangulation_type() = libMesh::TriangulatorInterface::PSLG;
-
-  // If we're using a user-requested subset of boundaries on that
-  // mesh, get their ids.
-  std::set<std::size_t> bdy_ids;
-
-  if (!xyd_opts.input_boundary_names.empty())
-  {
-    if (!xyd_opts.input_subdomain_names.empty())
-      mg.paramError(
-          "input_subdomain_names",
-          "input_boundary_names and input_subdomain_names cannot both specify an outer boundary.");
-
-    for (const auto & name : xyd_opts.input_boundary_names)
-    {
-      auto bcid = MooseMeshUtils::getBoundaryID(name, *mesh);
-      if (bcid == BoundaryInfo::invalid_id)
-        mg.paramError("input_boundary_names", name, " is not a boundary name in the input mesh");
-
-      bdy_ids.insert(bcid);
-    }
-  }
-
-  if (!xyd_opts.input_subdomain_names.empty())
-  {
-    // Make sure subdomain info caches are up to date
-    if (!mesh->preparation().has_cached_elem_data)
-      mesh->cache_elem_data();
-
-    const auto subdomain_ids =
-        MooseMeshUtils::getSubdomainIDs(*mesh, xyd_opts.input_subdomain_names);
-
-    // Check that the requested subdomains exist in the mesh
-    std::set<SubdomainID> subdomains;
-    mesh->subdomain_ids(subdomains);
-
-    for (auto i : index_range(subdomain_ids))
-    {
-      if (subdomain_ids[i] == Moose::INVALID_BLOCK_ID || !subdomains.count(subdomain_ids[i]))
-        mg.paramError("input_subdomain_names",
-                      xyd_opts.input_subdomain_names[i],
-                      " was not found in the boundary mesh");
-
-      bdy_ids.insert(subdomain_ids[i]);
-    }
-  }
-
-  if (!bdy_ids.empty())
-    poly2tri.set_outer_boundary_ids(bdy_ids);
-
-  poly2tri.set_interpolate_boundary_points(xyd_opts.add_nodes_per_boundary_segment);
-  poly2tri.set_refine_boundary_allowed(xyd_opts.refine_bdy);
-  poly2tri.set_verify_hole_boundaries(xyd_opts.verify_holes);
-
-  poly2tri.desired_area() = xyd_opts.desired_area;
-  poly2tri.minimum_angle() = 0; // Not yet supported
-  poly2tri.smooth_after_generating() = xyd_opts.smooth_tri;
-
-  std::vector<libMesh::TriangulatorInterface::MeshedHole> meshed_holes;
-  std::vector<libMesh::TriangulatorInterface::Hole *> triangulator_hole_ptrs(hole_meshes.size());
-  // This tells us the element orders of the hole meshes
-  // For the boundary meshes, it can be access through poly2tri.segment_midpoints.
-  std::vector<bool> holes_with_midpoints(hole_meshes.size());
-  bool stitch_second_order_holes(false);
-
-  // Make sure pointers here aren't invalidated by a resize
-  meshed_holes.reserve(hole_meshes.size());
-  for (auto hole_i : index_range(hole_meshes))
-  {
-    if (!hole_meshes[hole_i]->is_prepared())
-      hole_meshes[hole_i]->prepare_for_use();
-    if (hole_i < xyd_opts.hole_boundary_id_filters.size() &&
-        !xyd_opts.hole_boundary_id_filters[hole_i].empty())
-      meshed_holes.emplace_back(*hole_meshes[hole_i], xyd_opts.hole_boundary_id_filters[hole_i]);
-    else
-      meshed_holes.emplace_back(*hole_meshes[hole_i]);
-    holes_with_midpoints[hole_i] = meshed_holes.back().n_midpoints();
-    stitch_second_order_holes =
-        xyd_opts.stitch_holes.empty()
-            ? false
-            : ((holes_with_midpoints[hole_i] && xyd_opts.stitch_holes[hole_i]) ||
-               stitch_second_order_holes);
-    if (hole_i < xyd_opts.refine_holes.size())
-      meshed_holes.back().set_refine_boundary_allowed(xyd_opts.refine_holes[hole_i]);
-
-    triangulator_hole_ptrs[hole_i] = &meshed_holes.back();
-  }
-  if (stitch_second_order_holes &&
-      (xyd_opts.tri_elem_type == "TRI3" || xyd_opts.tri_elem_type == "DEFAULT"))
-    mg.paramError(
-        "tri_element_type",
-        "Cannot use first order elements with stitched quadratic element holes. Please try "
-        "to specify a higher-order tri_element_type or reduce the order of the hole inputs.");
-
-  if (!triangulator_hole_ptrs.empty())
-    poly2tri.attach_hole_list(&triangulator_hole_ptrs);
-
-  if (xyd_opts.desired_area_func != "")
-  {
-    // poly2tri will clone this so it's fine going out of scope
-    libMesh::ParsedFunction<Real> area_func{xyd_opts.desired_area_func};
-    poly2tri.set_desired_area_function(&area_func);
-  }
-  else if (xyd_opts.use_auto_area_func)
-  {
-    poly2tri.set_auto_area_function(
-        mg.comm(),
-        xyd_opts.auto_area_function_num_points,
-        xyd_opts.auto_area_function_power,
-        xyd_opts.auto_area_func_default_size > 0.0 ? xyd_opts.auto_area_func_default_size : 0.0,
-        xyd_opts.auto_area_func_default_size_dist > 0.0 ? xyd_opts.auto_area_func_default_size_dist
-                                                        : -1.0);
-  }
-
-  if (xyd_opts.tri_elem_type == "TRI6")
-    poly2tri.elem_type() = libMesh::ElemType::TRI6;
-  else if (xyd_opts.tri_elem_type == "TRI7")
-    poly2tri.elem_type() = libMesh::ElemType::TRI7;
-  // Add interior points before triangulating. Only points inside the boundaries
-  // will be meshed.
-  for (const auto & point : xyd_opts.interior_points)
-    mesh->add_point(point);
-
-  poly2tri.triangulate();
-
-  SubdomainID output_subdomain_id =
-      xyd_opts.has_output_subdomain_id ? xyd_opts.output_subdomain_id : 0;
-
-  if (xyd_opts.has_output_subdomain_name)
-  {
-    auto id = MooseMeshUtils::getSubdomainID(xyd_opts.output_subdomain_name, *mesh);
-
-    if (id == Elem::invalid_subdomain_id)
-    {
-      if (!xyd_opts.has_output_subdomain_id)
-      {
-        // We'll probably need to make a new ID, then
-        output_subdomain_id = MooseMeshUtils::getNextFreeSubdomainID(*mesh);
-
-        // But check the hole meshes for our output subdomain name too
-        for (auto & hole_ptr : hole_meshes)
-        {
-          auto possible_sbdid =
-              MooseMeshUtils::getSubdomainID(xyd_opts.output_subdomain_name, *hole_ptr);
-          // Huh, it was in one of them
-          if (possible_sbdid != Elem::invalid_subdomain_id)
-          {
-            output_subdomain_id = possible_sbdid;
-            break;
-          }
-          output_subdomain_id =
-              std::max(output_subdomain_id, MooseMeshUtils::getNextFreeSubdomainID(*hole_ptr));
-        }
-      }
-    }
-    else
-    {
-      if (xyd_opts.has_output_subdomain_id)
-      {
-        if (id != output_subdomain_id)
-          mg.paramError("output_subdomain_name",
-                        "name has been used by the input meshes and the corresponding id is not "
-                        "equal to 'output_subdomain_id'");
-      }
-      else
-        output_subdomain_id = id;
-    }
-    // We do not want to set an empty subdomain name
-    if (xyd_opts.output_subdomain_name.size())
-      mesh->subdomain_name(output_subdomain_id) = xyd_opts.output_subdomain_name;
-  }
-
-  if (xyd_opts.smooth_tri || output_subdomain_id)
-    for (auto elem : mesh->element_ptr_range())
-    {
-      mooseAssert(elem->type() == (xyd_opts.tri_elem_type == "TRI6"
-                                       ? TRI6
-                                       : (xyd_opts.tri_elem_type == "TRI7" ? TRI7 : TRI3)),
-                  "Unexpected element type " << libMesh::Utility::enum_to_string(elem->type())
-                                             << " found in triangulation");
-
-      elem->subdomain_id() = output_subdomain_id;
-
-      // I do not trust Laplacian mesh smoothing not to invert
-      // elements near reentrant corners.  Eventually we'll add better
-      // smoothing options, but even those might have failure cases.
-      // Better to always do extra tests here than to ever let users
-      // try to run on a degenerate mesh.
-      if (xyd_opts.smooth_tri)
-      {
-        auto cross_prod = (elem->point(1) - elem->point(0)).cross(elem->point(2) - elem->point(0));
-
-        if (cross_prod(2) <= 0)
-          mooseError("Inverted element found in triangulation.\n"
-                     "Laplacian smoothing can create these at reentrant corners; disable it?");
-      }
-    }
-
-  // The hole meshes are specified by the user, so they could have any
-  // BCID or no BCID or any combination of BCIDs on their outer
-  // boundary, so we'll have to set our own BCID to use for stitching
-  // there.  We'll need to check all the holes for used BCIDs, if we
-  // want to pick a new ID on hole N that doesn't conflict with any
-  // IDs on hole M < N (or with the IDs on the new triangulation)
-
-  // The new triangulation by default assigns BCID i+1 to hole i ...
-  // but we can't even use this for mesh stitching, because we can't
-  // be sure it isn't also already in use on the hole's mesh and so we
-  // won't be able to safely clear it afterwards.
-  const boundary_id_type end_bcid = hole_meshes.size() + 1;
-
-  // If a hole has its boundary layer mesh, we need to move the hole bcid to the "real" hole
-  // boundary in the boundary layer mesh. So we need to record them here.
-  std::vector<BoundaryID> hole_boundary_rec(hole_meshes.size());
-  std::iota(hole_boundary_rec.begin(), hole_boundary_rec.end(), 1);
-
-  // For the hole meshes that need to be stitched, we would like to make sure the hole boundary ids
-  // and output boundary id are not conflicting with the existing boundary ids of the hole meshes to
-  // be stitched.
-  BoundaryID free_boundary_id = 0;
-  if (xyd_opts.stitch_holes.size())
-  {
-    for (auto hole_i : index_range(hole_meshes))
-    {
-      if (xyd_opts.stitch_holes[hole_i])
-      {
-        free_boundary_id =
-            std::max(free_boundary_id, MooseMeshUtils::getNextFreeBoundaryID(*hole_meshes[hole_i]));
-        hole_meshes[hole_i]->comm().max(free_boundary_id);
-      }
-    }
-    for (auto h : index_range(hole_meshes))
-    {
-      libMesh::MeshTools::Modification::change_boundary_id(*mesh, h + 1, h + 1 + free_boundary_id);
-      hole_boundary_rec[h] = h + 1 + free_boundary_id;
-    }
-  }
-  boundary_id_type new_hole_bcid = end_bcid + free_boundary_id;
-
-  // We might be overriding the default bcid numbers.  We have to be
-  // careful about how we renumber, though.  We pick unused temporary
-  // numbers because e.g. "0->2, 2->0" is impossible to do
-  // sequentially, but "0->N, 2->N+2, N->2, N+2->0" works.
-  libMesh::MeshTools::Modification::change_boundary_id(
-      *mesh, 0, (xyd_opts.has_output_boundary ? end_bcid : 0) + free_boundary_id);
-
-  if (!xyd_opts.hole_boundaries.empty())
-  {
-    auto hole_boundary_ids = MooseMeshUtils::getBoundaryIDs(*mesh, xyd_opts.hole_boundaries, true);
-
-    for (auto h : index_range(hole_meshes))
-      libMesh::MeshTools::Modification::change_boundary_id(
-          *mesh, h + 1 + free_boundary_id, h + 1 + free_boundary_id + end_bcid);
-
-    for (auto h : index_range(hole_meshes))
-    {
-      libMesh::MeshTools::Modification::change_boundary_id(
-          *mesh, h + 1 + free_boundary_id + end_bcid, hole_boundary_ids[h]);
-      hole_boundary_rec[h] = hole_boundary_ids[h];
-      mesh->get_boundary_info().sideset_name(hole_boundary_ids[h]) = xyd_opts.hole_boundaries[h];
-      new_hole_bcid = std::max(new_hole_bcid, boundary_id_type(hole_boundary_ids[h] + 1));
-    }
-  }
-
-  if (xyd_opts.has_output_boundary)
-  {
-    const std::vector<BoundaryID> output_boundary_id =
-        MooseMeshUtils::getBoundaryIDs(*mesh, {xyd_opts.output_boundary}, true);
-
-    libMesh::MeshTools::Modification::change_boundary_id(
-        *mesh, end_bcid + free_boundary_id, output_boundary_id[0]);
-    mesh->get_boundary_info().sideset_name(output_boundary_id[0]) = xyd_opts.output_boundary;
-
-    new_hole_bcid = std::max(new_hole_bcid, boundary_id_type(output_boundary_id[0] + 1));
-  }
-
-  bool doing_stitching = false;
-
-  for (auto hole_i : index_range(hole_meshes))
-  {
-    const MeshBase & hole_mesh = *hole_meshes[hole_i];
-    auto & hole_boundary_info = hole_mesh.get_boundary_info();
-    const std::set<boundary_id_type> & local_hole_bcids = hole_boundary_info.get_boundary_ids();
-
-    if (!local_hole_bcids.empty())
-      new_hole_bcid = std::max(new_hole_bcid, boundary_id_type(*local_hole_bcids.rbegin() + 1));
-    hole_mesh.comm().max(new_hole_bcid);
-
-    if (hole_i < xyd_opts.stitch_holes.size() && xyd_opts.stitch_holes[hole_i])
-      doing_stitching = true;
-  }
-
-  const boundary_id_type inner_bcid = new_hole_bcid + 1;
-
-  // libMesh mesh stitching still requires a serialized mesh, and it's
-  // cheaper to do that once than to do it once-per-hole
-  libMesh::MeshSerializer serial(*mesh, doing_stitching);
-
-  // Define a reference map variable for subdomain map
-  auto & main_subdomain_map = mesh->set_subdomain_name_map();
-  for (auto hole_i : index_range(hole_meshes))
-  {
-    if (hole_i < xyd_opts.stitch_holes.size() && xyd_opts.stitch_holes[hole_i])
-    {
-      UnstructuredMesh & hole_mesh = dynamic_cast<UnstructuredMesh &>(*hole_meshes[hole_i]);
-      // increase hole mesh order if the triangulation mesh has higher order
-      if (!holes_with_midpoints[hole_i])
-      {
-        if (xyd_opts.tri_elem_type == "TRI6")
-          hole_mesh.all_second_order();
-        else if (xyd_opts.tri_elem_type == "TRI7")
-          hole_mesh.all_complete_order();
-      }
-      auto & hole_boundary_info = hole_mesh.get_boundary_info();
-
-      // Our algorithm here requires a serialized Mesh.  To avoid
-      // redundant serialization and deserialization (libMesh
-      // MeshedHole and stitch_meshes still also require
-      // serialization) we'll do the serialization up front.
-      libMesh::MeshSerializer serial_hole(hole_mesh);
-
-      // It would have been nicer for MeshedHole to add the BCID
-      // itself, but we want MeshedHole to work with a const mesh.
-      // We'll still use MeshedHole, for its code distinguishing
-      // outer boundaries from inner boundaries on a
-      // hole-with-holes.
-      const auto & hole_bdy_id_filter = (hole_i < xyd_opts.hole_boundary_id_filters.size())
-                                            ? xyd_opts.hole_boundary_id_filters[hole_i]
-                                            : std::set<std::size_t>();
-      libMesh::TriangulatorInterface::MeshedHole mh{hole_mesh, hole_bdy_id_filter};
-
-      // We have to translate from MeshedHole points to mesh
-      // sides.
-      std::unordered_map<Point, Point> next_hole_boundary_point;
-      const int np = mh.n_points();
-      for (auto pi : make_range(1, np))
-        next_hole_boundary_point[mh.point(pi - 1)] = mh.point(pi);
-      next_hole_boundary_point[mh.point(np - 1)] = mh.point(0);
-
-#ifndef NDEBUG
-      int found_hole_sides = 0;
-#endif
-      for (auto elem : hole_mesh.element_ptr_range())
-      {
-        if (elem->dim() != 2)
-          mooseError("Non 2-D element found in hole; stitching is not supported.");
-
-        auto ns = elem->n_sides();
-        for (auto s : make_range(ns))
-        {
-          auto it_s = next_hole_boundary_point.find(elem->point(s));
-          if (it_s != next_hole_boundary_point.end())
-            if (it_s->second == elem->point((s + 1) % ns))
-            {
-              hole_boundary_info.add_side(elem, s, new_hole_bcid);
-#ifndef NDEBUG
-              ++found_hole_sides;
-#endif
-            }
-        }
-      }
-      mooseAssert(found_hole_sides == np, "Failed to find full outer boundary of meshed hole");
-
-      auto & mesh_boundary_info = mesh->get_boundary_info();
-#ifndef NDEBUG
-      int found_inner_sides = 0;
-#endif
-      for (auto elem : mesh->element_ptr_range())
-      {
-        auto ns = elem->n_sides();
-        for (auto s : make_range(ns))
-        {
-          auto it_s = next_hole_boundary_point.find(elem->point((s + 1) % ns));
-          if (it_s != next_hole_boundary_point.end())
-            if (it_s->second == elem->point(s))
-            {
-              mesh_boundary_info.add_side(elem, s, inner_bcid);
-#ifndef NDEBUG
-              ++found_inner_sides;
-#endif
-            }
-        }
-      }
-      mooseAssert(found_inner_sides == np, "Failed to find full boundary around meshed hole");
-
-      // Retrieve subdomain name map from the mesh to be stitched and insert it into the main
-      // subdomain map
-      const auto & increment_subdomain_map = hole_mesh.get_subdomain_name_map();
-      main_subdomain_map.insert(increment_subdomain_map.begin(), increment_subdomain_map.end());
-
-      // We do not need the hole_bdy_id_filter anymore
-      for (const auto & bcid : hole_bdy_id_filter)
-        hole_boundary_info.remove_id(bcid);
-      // If we are stitching a hole boundary layer mesh, we need to reassign the bcid
-      if (hole_bdy_id_filter.size())
-      {
-        MooseMeshUtils::changeBoundaryId(
-            hole_mesh,
-            xyd_opts.hole_boundary_inner_id_defaults[hole_i].empty()
-                ? 1
-                : *xyd_opts.hole_boundary_inner_id_defaults[hole_i].begin(),
-            hole_boundary_rec[hole_i],
-            true);
-        hole_mesh.get_boundary_info().sideset_name(hole_boundary_rec[hole_i]) =
-            mesh->get_boundary_info().sideset_name(hole_boundary_rec[hole_i]);
-        mesh->get_boundary_info().remove_id(hole_boundary_rec[hole_i]);
-      }
-
-      mesh->stitch_meshes(hole_mesh,
-                          inner_bcid,
-                          new_hole_bcid,
-                          TOLERANCE,
-                          /*clear_stitched_bcids*/ true,
-                          xyd_opts.verbose_stitching,
-                          xyd_opts.use_binary_search);
-    }
-  }
-  // Check if one SubdomainName is shared by more than one subdomain ids
-  std::set<SubdomainName> main_subdomain_map_name_list;
-  for (auto const & id_name_pair : main_subdomain_map)
-    main_subdomain_map_name_list.emplace(id_name_pair.second);
-  if (main_subdomain_map.size() != main_subdomain_map_name_list.size())
-    mg.paramError("holes", "The hole meshes contain subdomain name maps with conflicts.");
-
-  mesh->unset_is_prepared();
-  return mesh;
-}
-
-std::vector<Point>
-generateOffsetPolyline(MeshGenerator * mg,
-                       std::unique_ptr<libMesh::UnstructuredMesh> & ply_mesh_u,
-                       std::vector<Point> & points,
-                       std::vector<Point> & mid_points,
-                       const bool outward,
-                       const Real thickness)
-{
-  // If the input points are empty, we will extract them from the input 1D mesh
-  if (points.empty())
-  {
-    mooseAssert(mid_points.empty(),
-                "If the input points are empty, the input mid_points must be also empty.");
-
-    TriangulatorInterface::MeshedHole bdry_mh(*ply_mesh_u);
-    MooseMeshUtils::collectKeyPointsFromMesh(bdry_mh, points, mid_points);
-  }
-  // Generate a very simple triangulation mesh so that we can get the outward normal vectors
-  libMesh::Poly2TriTriangulator poly2tri(*ply_mesh_u);
-  poly2tri.triangulation_type() = libMesh::TriangulatorInterface::PSLG;
-
-  poly2tri.set_interpolate_boundary_points(0);
-  poly2tri.set_refine_boundary_allowed(false);
-  poly2tri.set_verify_hole_boundaries(false);
-  poly2tri.desired_area() = 0;
-  poly2tri.minimum_angle() = 0; // Not yet supported
-  poly2tri.smooth_after_generating() = false;
-  if (mid_points.size())
-    poly2tri.elem_type() = libMesh::ElemType::TRI6;
-  poly2tri.triangulate();
-
-  // We need to serialize the mesh for next steps
-  libMesh::MeshSerializer serial(*ply_mesh_u);
-  // The mesh now only contains one side set that corresponds to the outer boundary with an ID of 0
-  auto bdry_list(ply_mesh_u->get_boundary_info().build_side_list());
-
-  // For each vertex, the shifting direction to form the offset is defined by the normal vectors of
-  // the two sides that contain the vertex.
-  // We gather the normals for each node on the boundary here, which are all vertices because of
-  // the pre-selection of nodes.
-  std::map<dof_id_type, std::vector<Point>> node_normal_map;
-  std::map<dof_id_type, Point> mid_node_normal_map;
-  for (const auto & bside : bdry_list)
-  {
-    const auto & side = ply_mesh_u->elem_ptr(std::get<0>(bside))->side_ptr(std::get<1>(bside));
-    // For linear elements, the side normal is constant and can be obtained straightforwardly;
-    // For quadratic elements, we need the normal at the shared vertices
-    const Point side_normal_0 =
-        mid_points.size()
-            ? getKeyNormal(ply_mesh_u->elem_ptr(std::get<0>(bside)), std::get<1>(bside), 0)
-            : ply_mesh_u->elem_ptr(std::get<0>(bside))
-                  ->side_vertex_average_normal(std::get<1>(bside));
-    const Point side_normal_1 =
-        mid_points.size()
-            ? getKeyNormal(ply_mesh_u->elem_ptr(std::get<0>(bside)), std::get<1>(bside), 1)
-            : side_normal_0;
-
-    if (node_normal_map.count(side->node_ptr(0)->id()))
-      node_normal_map[side->node_ptr(0)->id()].push_back(side_normal_0);
-    else
-      node_normal_map[side->node_ptr(0)->id()] = {side_normal_0};
-    if (node_normal_map.count(side->node_ptr(1)->id()))
-      node_normal_map[side->node_ptr(1)->id()].push_back(side_normal_1);
-    else
-      node_normal_map[side->node_ptr(1)->id()] = {side_normal_1};
-
-    if (mid_points.size())
-    {
-      const Point mid_node_normal =
-          getKeyNormal(ply_mesh_u->elem_ptr(std::get<0>(bside)), std::get<1>(bside), 2);
-      mid_node_normal_map
-          [ply_mesh_u->elem_ptr(std::get<0>(bside))->node_ptr(std::get<1>(bside) + 3)->id()] =
-              mid_node_normal;
-    }
-  }
-
-  std::vector<Point> mod_reduced_pts_list(points);
-  for (const auto & [node_id, normal_vecs] : node_normal_map)
-  {
-    mooseAssert(normal_vecs.size() == 2,
-                "Each vertex should be connected to exactly two sides in a polygon.");
-
-    const Point original_pt = *(ply_mesh_u->node_ptr(node_id));
-    // Form an average normal at the vertex from the two connected sides' normals
-    const Point move_dir =
-        (normal_vecs.front() + normal_vecs.back()).unit() * (outward ? 1.0 : -1.0);
-    // Consider four points of interest to determine the moving distance
-    // 1. the vertex point
-    // 2. point along normal_vecs.front() from the vertex with a distance thickness
-    // 3. point along normal_vecs.back() from the vertex with a distance thickness
-    // 4. point along move_dir from the vertex with a distance mov_dist
-    // The four points form a kite shape with its symmetry axis along 1-4 direction
-    // Angle 1-2-4 and angle 1-3-4 are right angles
-    // Angle 2-1-3 (i.e., theta) can be calculated using the interior product of
-    // v1 = normal_vecs.front() and v2 = normal_vecs.back():
-    // cos(theta) = (v1 . v2) / (|v1| * |v2|)
-    // Because of the right angles, the distance between 1 and 4 can be calculated as:
-    // mov_dist = thickness / cos(theta/2)
-    // and cos(theta/2) = sqrt((1 + cos(theta)) / 2)
-    const Real mov_dist =
-        thickness / std::sqrt((1.0 + (normal_vecs.front() * normal_vecs.back()) /
-                                         (normal_vecs.front().norm() * normal_vecs.back().norm())) /
-                              2.0);
-    mooseAssert(std::count(points.begin(), points.end(), original_pt) == 1,
-                "The original point should be found exactly once in the reduced points list.");
-    mod_reduced_pts_list[std::distance(points.begin(),
-                                       std::find(points.begin(), points.end(), original_pt))] =
-        original_pt + move_dir * mov_dist;
-  }
-
-  // To ensure no overlapping, we need to check set of four points
-  // p1 and p2 should be a pair of points before and after shifting
-  // p3 and p4 should be a pair of adjacent shifted points that are neither not p2
-  for (const auto & i_node_1 : make_range(mod_reduced_pts_list.size()))
-  {
-    const Point & p1 = points[i_node_1];
-    const Point & p2 = mod_reduced_pts_list[i_node_1];
-    for (const auto & i_node_2 : make_range(mod_reduced_pts_list.size()))
-    {
-      if (i_node_2 == i_node_1 || (i_node_2 + 1) % mod_reduced_pts_list.size() == i_node_1)
-        continue;
-      const Point & p3 = mod_reduced_pts_list[i_node_2];
-      const Point & p4 = mod_reduced_pts_list[(i_node_2 + 1) % mod_reduced_pts_list.size()];
-      if (thickness > 0)
-        if (geom_utils::segmentsIntersect(p1, p2, p3, p4))
-          mg->paramError(
-              "thickness",
-              "The thickness is so large that the mesh is tangled because the offset nodes "
-              "are no longer in the same order when following the original boundary. Please "
-              "reduce the thickness value.");
-    }
-  }
-
-  std::vector<Point> mid_mod_reduced_pts_list(mid_points.size());
-  for (const auto & [node_id, normal_vec] : mid_node_normal_map)
-  {
-    const Point original_pt = *(ply_mesh_u->node_ptr(node_id));
-    const Point move_dir = normal_vec.unit() * (outward ? 1.0 : -1.0);
-    mid_mod_reduced_pts_list[std::distance(
-        mid_points.begin(), std::find(mid_points.begin(), mid_points.end(), original_pt))] =
-        original_pt + move_dir * thickness;
-  }
-
-  // combine mod_reduced_pts_list and mid_mod_reduced_pts_list to get the final list of points for
-  // the layer mesh
-  std::vector<Point> layer_pts_list(mod_reduced_pts_list);
-  layer_pts_list.insert(
-      layer_pts_list.end(), mid_mod_reduced_pts_list.begin(), mid_mod_reduced_pts_list.end());
-
-  return layer_pts_list;
-}
-
-void
-collectKeyPointsFromMesh(libMesh::TriangulatorInterface::MeshedHole & bdry_mh,
-                         std::vector<Point> & points,
-                         std::vector<Point> & mid_points,
-                         const bool skip_node_reduction)
-{
-  for (const auto i : make_range(bdry_mh.n_points()))
-  {
-    if (skip_node_reduction || !geom_utils::arePointsColinear(
-                                   bdry_mh.point((i - 1 + bdry_mh.n_points()) % bdry_mh.n_points()),
-                                   bdry_mh.point(i),
-                                   bdry_mh.point((i + 1) % bdry_mh.n_points())))
-    {
-      points.push_back(bdry_mh.point(i));
-      if (bdry_mh.n_midpoints() == 1 && skip_node_reduction)
-        mid_points.push_back(bdry_mh.midpoint(0, i));
-    }
-  }
-}
-
-Point
-getKeyNormal(const Elem * elem, const unsigned int s, const unsigned int key_node_index)
-{
-  const std::unique_ptr<const Elem> face = elem->build_side_ptr(s);
-  mooseAssert(face->type() == ElemType::EDGE3,
-              "Only elements with EDGE3 sides are supported in this function.");
-  std::unique_ptr<libMesh::FEBase> fe(
-      libMesh::FEBase::build(2, libMesh::FEType(elem->default_order())));
-  const std::vector<Point> & normals = fe->get_normals();
-  std::vector<Point> ref_pts = {face->reference_elem()->point(key_node_index)};
-  fe->reinit(elem, s, TOLERANCE, &ref_pts);
-  return normals[0];
 }
 }
